moving validator instantiation to the submitter configuration so it is create along side the subscriptionValidatorInterceptor
