+++ hapi-fhir-jpaserver-subscription/src/main/java/ca/uhn/fhir/jpa/subscription/config/SubscriptionConfig.java
@@ -20,20 +20,10 @@
 package ca.uhn.fhir.jpa.subscription.config;
 
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
-import ca.uhn.fhir.jpa.model.config.SubscriptionSettings;
 import ca.uhn.fhir.jpa.subscription.match.matcher.matching.SubscriptionStrategyEvaluator;
-import ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.IChannelTypeValidator;
-import ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.RegexEndpointUrlValidationStrategy;
-import ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.RestHookChannelValidator;
-import ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.SubscriptionChannelTypeValidatorFactory;
 import ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.SubscriptionQueryValidator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
-
-import static ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.RestHookChannelValidator.IEndpointUrlValidationStrategy;
-import static ca.uhn.fhir.jpa.subscription.submit.interceptor.validator.RestHookChannelValidator.noOpEndpointUrlValidationStrategy;
 
 @Configuration
 public class SubscriptionConfig {
@@ -41,23 +31,5 @@ public class SubscriptionConfig {
 	public SubscriptionQueryValidator subscriptionQueryValidator(
 			DaoRegistry theDaoRegistry, SubscriptionStrategyEvaluator theSubscriptionStrategyEvaluator) {
 		return new SubscriptionQueryValidator(theDaoRegistry, theSubscriptionStrategyEvaluator);
-	}
-
-	@Bean
-	public IChannelTypeValidator restHookChannelValidator(SubscriptionSettings theSubscriptionSettings) {
-		IEndpointUrlValidationStrategy iEndpointUrlValidationStrategy = noOpEndpointUrlValidationStrategy;
-
-		if (theSubscriptionSettings.hasRestHookEndpointUrlValidationRegex()) {
-			String endpointUrlValidationRegex = theSubscriptionSettings.getRestHookEndpointUrlValidationRegex();
-			iEndpointUrlValidationStrategy = new RegexEndpointUrlValidationStrategy(endpointUrlValidationRegex);
-		}
-
-		return new RestHookChannelValidator(iEndpointUrlValidationStrategy);
-	}
-
-	@Bean
-	public SubscriptionChannelTypeValidatorFactory subscriptionChannelTypeValidatorFactory(
-			List<IChannelTypeValidator> theValidorList) {
-		return new SubscriptionChannelTypeValidatorFactory(theValidorList);
 	}
 }
